set cards-on-the-board array at the main page to share between two components
transfer function of adding cards to deck component;\n remove commented code
