+++ src/components/ui/Deck/Deck.tsx
@@ -1,12 +1,13 @@
+import { FC } from "react";
 import { v4 as uuidv4 } from "uuid";
 
 import { NORTHERN_REALMS_CARDS_ARRAY, NORTHERN_REALMS_CARDS_QUANTITY } from "../../../constants";
 import { useCardSetup } from "../../../hooks";
-// import { CardProps } from "../../../types";
+import { CardsOnBoardUpdater, CardType } from "../../../types";
 import { Card } from "../Card";
 import { CardRow } from "../CardRow";
 
-export const Deck = () => {
+export const Deck: FC<CardsOnBoardUpdater> = ({ setCardsOnBoard }) => {
   const { cardsInDeck, loading, error } = useCardSetup(
     NORTHERN_REALMS_CARDS_ARRAY,
     NORTHERN_REALMS_CARDS_QUANTITY
@@ -14,8 +15,44 @@ export const Deck = () => {
 
   if (loading) return <p>Loading...</p>;
 
+  const addCardToBoard = (selectedRowType: CardType) =>
+    setCardsOnBoard((prevState) =>
+      prevState.map((row) =>
+        row.type === selectedRowType
+          ? {
+              ...row,
+              cards: [
+                ...row.cards,
+                {
+                  name: `${Math.floor(Math.random() * 100)}`,
+                  description: "card!",
+                  type: "close",
+                  points: Math.floor(Math.random() * 10)
+                }
+              ]
+            }
+          : row
+      )
+    );
+
   return (
     <div style={{ width: "100%" }}>
+      <button
+        style={{ marginTop: "20px", marginRight: "20px", padding: "10px" }}
+        onClick={() => addCardToBoard("close")}>
+        Add close fight card to the board
+      </button>
+      <button
+        style={{ marginTop: "20px", marginRight: "20px", padding: "10px" }}
+        onClick={() => addCardToBoard("range")}>
+        Add range card to the board
+      </button>
+      <button
+        style={{ marginTop: "20px", padding: "10px" }}
+        onClick={() => addCardToBoard("siege")}>
+        Add siege card to the board
+      </button>
+
       <h3>{error}</h3>
       <h2>Deck</h2>
       <CardRow type="deck">
@@ -23,38 +60,6 @@ export const Deck = () => {
           <Card name={name} description={description} type={type} points={points} key={uuidv4()} />
         ))}
       </CardRow>
-
-      {/*{renderRestOfTheCards(availableCards)}*/}
     </div>
   );
 };
-
-// probably useless component, todo: remove later
-// now it's just to make code more readable
-// function renderRestOfTheCards(cards: CardProps[]) {
-//   return (
-//     <>
-//       <h2 style={{ marginTop: "30px" }}>Rest</h2>
-//       {cards.length >= 1 && (
-//         <div style={{ position: "relative", width: "200px" }}>
-//           {renderFirstCard(cards[0], 0, 0)}
-//           {cards.length >= 2 && renderFirstCard(cards[0], -1, -5)}
-//           {cards.length >= 3 && renderFirstCard(cards[0], -2, -10)}
-//         </div>
-//       )}
-//     </>
-//   );
-// }
-
-// probably useless component, todo: remove later
-// now it's just to make code more readable
-// function renderFirstCard(firstCard: CardProps, zIndex: number, positionRight: number) {
-//   const { name, description, type, points } = firstCard;
-//
-//   return (
-//     <div
-//       style={{ position: "absolute", zIndex: `${zIndex}`, top: "0", right: `${positionRight}%` }}>
-//       <Card name={name} description={description} type={type} points={points} key={uuidv4()} />
-//     </div>
-//   );
-// }
