+++ src/types/components/ui/PlayerBoard.ts
@@ -1,6 +1,16 @@
+import React from "react";
+
 import { CardProps, CardType } from "./Card";
 
 export type RowProps = {
   type: CardType;
   cards: CardProps[];
+};
+
+export type CardsOnBoardArray = {
+  cardsOnBoard: RowProps[];
+};
+
+export type CardsOnBoardUpdater = {
+  setCardsOnBoard: React.Dispatch<React.SetStateAction<RowProps[]>>;
 };
